requirements.txt
redis==2.10.6
requests
SQLALchemy==1.2.6
mysql-connector-python-rf==2.1.3
pymongo==3.5.1
futures
Pillow==5.1.0
PyExecJS==1.5.0
python-Levenshtein
pathlib==1.0.1
demjson==2.2.4
pytesseract==0.2.0
user-agent==0.1.9
#cv
#sudo apt install tesseract-ocr  #3.04.01-4
mmh3==2.5.1
bitarray>=0.3.4
jieba>=0.39
#pip install setuptools cffi 'cython>=0.28' git+git://github.com/gevent/gevent.git#egg=gevent

Flask==0.12.2     
Flask-Script==2.0.6      
matplotlib==2.2.2      
mock==2.0.0      
numpy==1.14.2     
pbr==4.0.1
Pillow==5.1.0
protobuf==3.5.2.post1
pyparsing==2.2.0      
python-dateutil==2.7.2      
pytz==2018.3     
redis==2.10.6     
setuptools==39.0.1     
six==1.11.0     
subprocess32==3.2.7      
#tensorboard==1.7.0      
#tensorflow==1.7.0      
#termcolor==1.1.0      
#uWSGI==2.0.17     
uWSGI
Werkzeug==0.14.1     
wheel==0.31.0